Exported FieldClickedFourTime
